requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.8
aiosignal==1.3.1
attrs==24.2.0
cffi==1.17.1
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.0
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
yarl==1.18.0
yt-dlp==2024.11.18
